feat: 유저의 비식별처리 히스토리 조회 로직 추가

+++ src/main/java/com/imcloud/saas_user/common/entity/LogDetail.java
@@ -31,7 +31,7 @@ public class LogDetail extends Timestamped {
     private String description;
 
     @Column(name = "data_size", nullable = false)
-    private int dataSize;
+    private Long dataSize;
 
     @ManyToOne(fetch = FetchType.LAZY)
     @JoinColumn(name="activity_id")
